refactor: add greeter sample and collaps Bau.Exec into one file
