+++ src/samples/Bau.Samples.Greeter/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ScriptCs.Contracts" version="0.9.0" targetFramework="net45" />
+  <package id="StyleCop.MSBuild" version="4.7.49.0" targetFramework="net45" developmentDependency="true" />
+</packages>
